content-service: grpc layer pagination mapper; protos: content entities changed

+++ services/content-service/src/main/java/dev/animedia/contentservice/domain/shared/model/Page.java
@@ -6,9 +6,9 @@ import java.util.function.Function;
 public record Page<T>(
     List<T> content,
     long totalElements,
-    long totalPages,
-    long pageNumber,
-    long pageSize,
+    int totalPages,
+    int pageNumber,
+    int pageSize,
     boolean isFirst,
     boolean isLast,
     boolean hasNext,
